Move database logic to services
